requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
flask-socketio==5.3.6
tensorflow==2.12.0
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
librosa==0.10.0
scikit-learn==1.2.2
deepface==0.0.75
mediapipe==0.10.14
